+++ src/components/Me/index.jsx
@@ -1,14 +1,37 @@
 import React from "react";
+import M from "materialize-css";
+
+import { GistContext } from "../../contexts/GistContext";
+import { storage } from "../../services/Firebase";
+
 import MyPicture from "./MyPicture";
 import MyInfo from "./MyInfo";
 
+const HandleAfterDocumentLoaded = () => {
+    const elems = document.querySelectorAll(".materialboxed");
+    M.Materialbox.init(elems, { inDuration: 500, outDuration: 500 });
+};
+
 const Me = () => {
+    const gists = React.useContext(GistContext);
+    const [imageUrl, setImageUrl] = React.useState(null);
+    const storageRef = storage.ref()
+    React.useEffect(() => {
+        storageRef.child('minlaxz-dev/images/itssssme.jpg').getDownloadURL()
+            .then((url) => {
+                setImageUrl(url);
+            })
+        HandleAfterDocumentLoaded();
+    }, [storageRef]);
+
     return (
-        <div className="row" style={{ marginTop: "3%" }}>
-            <MyPicture />
-            <MyInfo />
+        <div className="container row" style={{ marginTop: "3%" }}>
+            <MyPicture imageUrl={imageUrl} />
+            <MyInfo gists={gists} />
         </div>
     );
 };
+
+
 
 export default Me;
